fix: extract errors from generation script
