+++ api/voice-react-native-sdk.api.md
@@ -123,6 +123,14 @@ namespace AuthorizationErrors {
         name: string;
         solutions: string[];
     }
+    class PayloadSizeExceededError extends TwilioError {
+        constructor(message: string);
+        causes: string[];
+        description: string;
+        explanation: string;
+        name: string;
+        solutions: string[];
+    }
     class RateExceededError extends TwilioError {
         constructor(message: string);
         causes: string[];
