+++ package.json
@@ -28,7 +28,7 @@
     "build:api:documenter": "api-documenter markdown --input-folder ./temp --output-folder ./docs/api",
     "build:constants": "node ./scripts/generate-constants.js",
     "build:docs": "yarn run prepare && yarn run build:api:extractor && yarn run build:api:documenter",
-    "build:errors": "node ./scripts/errors.js",
+    "build:errors": "node ./scripts/generate-errors.js",
     "check:lint": "eslint \"**/*.{js,ts,tsx}\"",
     "check:test": "jest",
     "check:type": "tsc --noEmit",
